Database defined hooks (#7)
Create a method to add hooks in the database. This allows hooks to be removed as well as added. And it removes the need to conform with naming conventions.

This will allow Phoenix to ship with certain behaviors preinstalled but allow other behaviors to be substituted by database calls. The current hook system requires that files be deleted to disable them. This makes updates more difficult, as it is hard to tell if a file should be added or if it was deliberately removed. Functionality using this form can simply delete the database reference. So the core hooks can be moved out of the file registration system and use this system instead.

This change is backwardly compatible. Any existing file-based hooks will continue to work. This allows store owners and App developers to deploy hooks as now. But the core code can instead use these instead of file-based hooks. This will leave the hooks directory to the customizers so that updates won't have to change it.

* New hooks table in install/oscommerce.sql

Add a table to define hooks.  This SQL needs to be run manually in updates.  

* Add loading from database in includes/classes/hooks.php

Alters the hooks class so that hooks can be loaded from either files or the database.  Requires the table be created first.

* Move hook code higher in includes/application_top.php

My goal is to make changes as high in the file as possible.  So I needed to move the hook code as early as possible.  Three lines (five with comment and blank) moved.  One added.  None deleted.

* Update copyright year in all three files

(cherry picked from commit a459afc6075373eb5271c1396302befa6ef8a3ae)