Admin autoloader
Updates the catalog autoloader functions to explicitly sort and creates a separate function for building the index of class files.

Adds an admin autoloader function.

Modifies the admin/includes/application_top.php to register the admin and catalog autoloaders in that order. So if a class exists in admin, it will use that one. If not, it will look in catalog.