Unify pipeline, page, and siteWide hooks
Creates a concept of pipeline hooks.  A pipeline being a group of pages with the same hook call.  

Modifies registration so that hooks can be registered to aliases.  

Registers pipeline and siteWide hooks to the current page.  So a siteWide hook will run for any invocation.  A pipeline hook will run for any invocation on a page that registers that pipeline (after the pipeline is registered).  A page hook will only trigger on the page that registers it.  So if someone wants to trigger a siteWide hook only on a particular page, they can set it up as a page hook.  

Change registers in the application_top.php files to use the new page registration.  

(cherry picked from commit 6231f9fc6e6c73420ec8a6173d270d0d0b84c934)