Prepare for switchable templates
In order to switch templates, we need to be able to map to a
separate template.  So we include cm_template.php to do the actual
include that we want.

Move all the HTML from the module file to the template file.

While there update the G Burton copyrights and licenses, as per a
separate conversation.  And extend abstract_executable_module to
reduce boilerplate.  And update to PHP 7 standards.