Prepare to deprecate admin order class
There is currently an error because the admin version of the order
class loads the country differently.  This updates the admin version
to match  the catalog version.

Longer term, the solution is probably to simply eliminate the admin
version of the class.  Then it will just use the catalog version and
we won't have to worry about this kind of inconsistency.  But to
avoid the problems that occur when people fail to delete the file,
just updating it for now.  We can revisit after the 1.0.6.0 release.