Replace session function calls
When osC was originally created, the session handling was rather
manual and it made sense to use helper functions.  Since then the
session handling has become much simpler.  It is now easier to use
the session superglobal directly.

This change replaces accesses of global references to the session
variables with the session variables themselves.  It also finishes
replacing the tep_session_register, tep_session_unregister, and
tep_session_is_registered calls in these files with direct accesses.

This is intended as one of multiple commits.  This set of files
relates to payments.