Create loginRequired pipeline
Adds a database hook for a loginRequired pipeline.
Adds a new function (for the hook to call).
Modifies the navigation history to work with them.

The new check (that $page['page'] is set) is better than the old,
as it checks the actual requirement.  The old only checked that it
was passed an array.  If the array was empty, it soldiered on.
This mattered because the hook passes an empty array by default.