Restore previous redirect behavior
Modifies the loginRequired declaration to restore the previous
behavior of returning to the payment page if the confirmation or
process page is hit without an active login session.

I do not know why this was the behavior previously.  But I do not
see an immediate need to change it.  This modifies the framework to
allow such an override and implements it for the two pages that
previously worked this way.