Fix password update on older format
When an older database is used with the newer code, a customer may
have a password using an older format.  We know how to validate these
and do so.  But afterward, we want to convert the stored password to
the newer, more secure format.  This takes a development version of
this capability and matches it to the form used to update the password
elsewhere.

(cherry picked from commit 5404d1442d4186a7254ee6302dadaef0a62e5f27)