Restrict customer data on create_account and admin/customers
This change allows people to select modules for only partial
installation.  So they don't display in account creation or the
admin customer edit form.  But do display somewhere else.

For updates:

UPDATE configuration
 SET configuration_value
       = CONCAT(configuration_value, ';create_account;customers')
 WHERE configuration_key LIKE 'MODULE_CUSTOMER_DATA%PAGES'

will give the current behavior for all installed modules.  Then
the configuration can be edited for any that need modified.
Presumably that will be easier if the current behavior is mostly
correct except for one or two modules.