Add tep_db_copy to copy a set of records
Sometimes, we want to make a copy of an existing set of records.
For example, if we want to make a duplicate copy of a product.

This allows an entire set of records to be copied at once, both
reading from and writing to the database.  This is more efficient
than reading the records separately from writing, because it makes
fewer round trips to the database and passes less information.