Convert to class
The specials expire function was called from application_top.  So
we want to move it to a system hook.  By converting to a class at
the same time, we allow it to be autoloaded rather than requiriing
a separate load step.

Also improved efficiency by combining the SELECT and UPDATE steps
into one, replacing a loop of expiration with a single SQL statement.