Switch application_top to use system hooks
This set of changes switches to use the new system hooks.  Most of
which were previously committed, but some are added here.

Also includes some changes to use renamed functions.

Note that some previous commits require this commit to function.
In particular, the migration of specials and whos_online to
system versioned classes.  They unfortunately could not be
combined into this commit without breaking git's history.

Note that this requires database changes.  Included in phoenix.sql
here, they would have to be run manually if upgraded.