Always add totals to order object
Previous versions only added the totals to the order object if it
was loaded from the database.  If loaded from the cart, the totals
were set in a separate variable just for that purpose.

This always adds the totals to the order object.  Which means that
we can use the same code to get the totals regardless of whether we
loaded the order from the cart or the database.

This fixes a bug where the totals were not appearing in IPN
generated order confirmation emails.  And it makes fewer global
variables.