Fix infobox forms
As per the HTML spec, forms can't start inside a table but outside
a TD or TH.  So any form that needs to cover multiple elements in a
table should be outside the table.  Otherwise Javascript won't be
able to access the form because it can't parse the DOM.

This fixes things like editing the Braintree module's configuration,
which relies on Javascript to set one of the values.

Left forms that start and end inside the same table element, as those
are fine.