Install at https: if accessed by https
If the store owner accesses the install with an https URL, then
install the store at https.  Otherwise assume they don't have https
installed and install at http.

Also simplifies the path check.

Allows stores to be installed in a path that includes install, e.g.
as a username or a subfolder.  Still can't install to a path that
includes install/install.php but hopefully people won't find that
limiting.