Deprecate USE_PCONNECT
Modernly, all USE_PCONNECT does is prepend p: to the DB_SERVER value.
It's just as easy to do that directly.  So for the few people who
use persistent connections, that is sufficient.

USE_PCONNECT was deprecated in catalog some time ago.  This catches
admin up to it.  And prepares for removing it from the configure.php
files.