New compatibility class for cookie handling
PHP 7.3 introduced a new method signature to handle the samesite
configuration option for cookies.  Prior to that, samesite was
handled by exploiting a bug in the options that allowed one to add
and arbitrary header to the cookie.  That bug was fixed in 7.3.

Provides two versions of a Cookie class with a common interface.
Which version is loaded depends on the PHP version.  If 7.3 or
higher, we use the new simple method signatures.  Before that, we
use the bug and pass the options in the URL manually.