Deprecate ENABLE_SSL/HTTPS
At one time, osCommerce used a mixture of http and https pages.  But
now CE Phoenix recommends only using HTTPS in live sites.  So there's
no point in having non-SSL configuration.  I.e. we always set the
HTTP and HTTPS to the same values.  So we only need one.

This simplifies by only using the HTTP values and not checking the
ENABLE_SSL setting.  So long as everything is configured as advised
this will make no change.