Rationalize configuration
There are a number of configuration values that are never actually
configured.  In practice, they are just used as constants.  So this
removes them from the configuration files, leaving only the values
that are actually configured to different values.

Eliminates the HTTP/HTTPS concept.  For some time, we have recommended
that store owners configure both HTTP_SERVER and HTTPS_SERVER to the
same value.  So why have separate configuration?  If anyone still
needs this kind of functionality, they can implement it as an App and
override href_link.

Move common configuration from code into configuration.  In
particular, error_reporting is moved into the configure.php file.
Also, the time zone setting is done directly in the file rather than
setting a configuration value and setting it elsewhere.

Moved configuration overrides and the auto-install code into the
default configure.php file.  That way people who need them will
have them by default.  But once installed, both disappear as
unnecessary (at least for most users).

Also improved the format of the configure.php files in the install
code so that it would be easier to read.  And updated from define to
const.