Move checkout modules to system/segments
When the checkout modules were created, the system/segments had not
yet been created.  So they were created in modules as the best of a
set of weak choices.  But now we do have system/segments and that
would be a better place for them.

Doing this change now so as to avoid compounding the problem by
adding more files in the wrong location.

Note that this breaks any payment module that accesses the checkout
modules.  I fixed the core ones as part of this commit.  But of
course not all payment modules are core.