Fix payment module handling
Finished the migration of checkout_confirmation to use the modular
checkout.

Changed the payment validation to use global rather than local
variables.  This fixes a bug when there were multiple payment methods
that prevented checkout getting past the payment page.

Adjusted the order of payment module initialization on the checkout
payment page to match the old order.  This fixed a bug where
disabling a module for all but specific zones was not working.

Changed the functions to static so as not to bother instantiating a
Checkout object.  Done at this time to facilitate making a call from
one method to another.