Restrict payment modules by billing zone
With the exception of the COD module (where the delivery address is
the de facto billing address), we want to disable payment modules
based on the zone to which they are billed.

This behavior had changed in 1.0.5.1 when the COD module's behavior
was given to all payment modules.  This reverts the effects of that
change.  The COD module has its own logic.  But all other core payment
modules now use the same logic.