Refactor and hook admin order listing
When an order status does not exist in the order_statuses table but
there are references to it in the orders table, it makes some orders
effectively invisible.  This changes so that it will always display
the order, regardless of status.

Some minor code tweaks while there.

Added new hookpoints to inject additional filter criteria or to
modify the order of the listing.