Allow non-zero flags
Originally this function was just copied from a payment module.
That module handled the case where it was installed outside Phoenix.
But we don't need to handle that case in Phoenix core.

This removes the check for the existence of the flag columns.

And this allows (but does not require) the caller to set values for
the flags.  By defaulting to zero, we maintain backwards compatibility
with callers that do not set the flags.  But callers that prefer
non-zero values can now pass them into the function.

https://github.com/gburton/CE-Phoenix/issues/1014