Consistent use_function in configuration editing
As noted in https://github.com/gburton/CE-Phoenix/issues/1013
sometimes we support static methods, sometimes only object methods,
sometimes we just throw an error.  This makes the four pages that
work with use_function values do the same thing.

Now we always check if the function is callable.  If not, we display
a warning.  We use the most generic check, so it basically only
checks if call_user_func would generate an error.  We make no
assumptions about what kind of function it is.  We do ensure that
if it is an object method call, that the object exists first.