Disable modules when zone requested but not set
Shipping and payment modules can be restricted by zone.

This had been checking if the zone was set at all.  If not, it was
always allowing the module.  This change modifies it to disable the
module whenever a zone is requested but not there.

This is more compatible with the older behavior, which implicitly
would not match an empty zone ID.  But now that is an explicit check.