Deprecate tep_guarantee functions in favor of Guarantor methods
The new Guarantor class obviates the need for tep_guarantee functions.
This replaces them with calls to the equivalent class methods.

This allows for the removal of the function definitions from the
hooks class.  So reducing the side effects of loading that class
and making it possible to do the guarantees without loading that
class.

This also fixes session hooks and delegates callback building to a
separate method for readability and simplicity.