Use email address class validation
Originally there was a validation function for email addresses.  The
customer data changes switched to a method on the cd_email class for
processing the email address input.  These two pages do not use the
customer data input or processing.  So they continued to use the old
way.

This will allow us to deprecate the functions/validations.php file.