Remove obsolete mode
We no longer mix SSL and non-SSL links.  So there is no need to
store the mode parameter nor pass it to tep_href_link.

Avoids notices when there is a login redirect during checkout.

For readability and robustness, I prefer $GLOBALS to using the
global keyword.  It is easier to read because it indicates that
the variable is the global instance when it is used rather than
at a separate location.  And since PHP implements the global
keyword by mapping global $variable to

$variable = &$GLOBALS['variable'];

it is less likely that someone will stomp on the reference if
the variable is accessed from $GLOBALS.  I.e. no one would
expect

$variable =& $something_else

to make the global variable a reference to $something_else
after the previous code, but that is the most natural reading
when using the global keyword.  It just doesn't work that way.

Anyway, where it doesn't make the code significantly longer, I
have favored using $GLOBALS to the global keyword.  In most
cases, this makes the code more compact as well.

In the future, I might even deprecate the global keywords
entirely in favor of the manual mapping.  But for now, I
continue to favor conciseness.