Avoid notices
When a customer data module is turned off, it leaves empty spots
in the data.  These create notices when access is attempted.

This change uses the get method to work around that.  It either
finds the data (if available) or suppresses the notice.

An alternative to this is to turn off E_NOTICE in error reporting.