Refactor What's New box and Product class
Revises the Product class to be easier to inject dependencies.
This includes the ability to lazy-load extended information.

Adds a new box template component to allow for easy editing of all
boxes that use it at once.  The block_template is updated to allow
box modules to call component templates directly.

Adds a new product card template component for a coherent display
of product information.  The component expects to be passed a
Product class object.  The component will also accept configuration
in a $card array.  The component does not require such configuration
if the default behavior is desired.

Refactors the What's New box to use the new components and classes.

Removes reliance on tep_random_select, which was a bit over-optimized
for the What's New use case.  That specialization is now moved into
the module.

Adds a capabilities_manager base class.  This is what powers the
lazy-loading logic, which is configured in the product_loader class.
The product_builder class contains static logic that does not involve
loading information from the database.  Logic for displaying
information stored in the Product class is in the class itself,
including three pricing functions.

The Product class does not know how to load its data.  It relies on
builder classes for that.  The first one is product_by_id, which
loads a product given a product_id.  This approach keeps the Product
class logic simpler while maximizing flexibility.