Better subquery handling
The original version starts from the beginning of the string and
works its way back.  This breaks when a subquery has the same
keywords as it tries to find.

This version will work if the subquery is in the FROM section.  It
searches backward from the back of the string rather than forward
from the front.  So if there are two WHERE clauses, it will find
the second one instead of the first.