Separate category tree display, data, and structure
The current category tree class mixes several things.  It creates a
tree structure to hold the category data.  It loads and exposes the
data.  And it has specific displays of the data.

This separates into multiple classes.  A generic tree structure with
tree operations.  A category tree implementation of that, which loads
and exposes the data.  A display class.  And a controller class for
the display.

This also simplifies accesses of the tree structure.  Both because
some could be simpler and through the addition of a parent_of data
structure.  This saves iterating to find a category in the tree.
We can simply look up the location of any category.  And we can
find its ancestors through iteration of a simple lookup.

The separation of data and display requires modification of uses of
the category tree.  In particular, the categories box.