Flip category depth logic
Previously, the logic was if there are products, then it's products
depth.  Otherwise, if there are categories, it's nested.  If there
are neither products nor categories, show as products depth, and
display a No Products message.

This flips the logic to check if there are categories first.  If so,
it is nested depth.  Otherwise, it is products depth.  This is
simpler and requires fewer database queries.  Usually no additional
queries, as the category tree would have already been built at this
point.  So we can just query the data structure.

This does the same thing except when there are both products and
subcategories in the same category.  Then this shows as nested
rather than products.  This hides the products in the default
install.  But a cm_in_product_listing.php could be made to handle
that.  That would show both categories and products at the same
time.