Reorder application start with template before language
The template must be loaded before the language file.  But the title
requires the language file to be loaded before the template file.

Resolves by setting the title after the initial load.

The template also needs to be loaded before the customer data
modules.  Which is now easy, as it gets loaded before the language.
So move the customer data and customer modules back to their previous
location.