Don't test array as string
This is not one of the places that tep_not_null should be converted to
Text::is_empty.

Rewriting to only accept valid currency configurations.  In particular,
the currency value must be defined, numeric, and not zero.  This will
implicitly check that it is not an empty array (the old check).