Add override template
Adds a template that, when enabled, will override files in the
default template with files that appear under the override template.

Note that this applies to both language and template files.  So
when editing these, the correct process will be

1.  Copy the original to the override template directory tree.
2.  Edit the file there while leaving the original alone.

This allows us to provide core implementations of these while still
allowing store owners to edit these files with custom behaviors
appropriate to their stores.