Load products without status limitation in admin
In the catalog, we don't need the data for inactive products that
we don't display anyway.  In admin, we do.  Because we want to be
able to edit inactive products.

This creates a separate method for admin so as to load the data
regardless of product status while retaining the original catalog
behavior.