Better naming for text input methods
Traditionally catalog and admin used a function called
tep_db_prepare_input to prepare user input to be put in the database.

In admin, this function simply trims off leading and trailing
whitespace after running stripslashes.

In catalog, this function also sanitizes user input by consolidating
multiple adjacent spaces into one and remove < and > signs.  This
prevents the user from entering HTML or more dangerously, Javascript.

We are moving to using autoloaded classes that, when the
functionality is the same, are used in both catalog and admin.  So
we will no longer have different versions of tep_db_prepare_input.
As a result, we do not want to use the same replacement for admin
inputs that may contain HTML.  We previously added a new method
for that circumstance.  But the name was not obvious.

Now, while the two names are used in a limited set of files,
switching to new, more rational names.

The catalog replacement will be called Text::input.  It is the
input equivalent of Text::output, sanitizing the input before
storing it.  It still removes multiple spaces and HTML.

The admin version will be Text::prepare.  It will simply strip any
slashes and trim whitespace.  It will be used for admin fields
that might contain HTML.  Most admin fields will use Text::input
as catalog does.

I note but do not act at this time on removing stripslashes as
unnecessary without gpc_magic_quotes.  Also, I think that input
could call prepare and then sanitize the result more efficiently
and without any change in behavior.  Because the current logic
consolidates multiple trailing spaces into one and then trims them.
It would be more efficient to trim first.

In the future, we will probably simplify these further.  Perhaps
Text::prepare is unnecessary and can just be trim.  But I do not
want to do that at this point in the 1.0.8.0 release cycle.