Mark unused tep_session_functions as deprecated
We are moving away from thin wrapper functions around PHP functions.
In many of these cases, the original function had logic in it.  But
as the PHP functions improved, the logic became unnecessary.
Further, we are moving away from function files in favor of class
files that can be autoloaded.  So eliminating unnecessary functions
serves two purposes.

Also makes the admin session functions use the underlying functions
themselves so as to make the functions unused.

And the tep_rand() function is itself a thin wrapper around mt_rand
when called with no arguments.

Deprecating these functions now means that they can be removed in
the 1.0.8.* series.