Allow deletion of categories without subcategories
If there are no subcategories, the category will have no children.
This causes notices, as the children list is unset in that case.

This modifies to check for that case by using get_children, which
already included an existence check.  So it will now properly
attempt to process an empty list and do nothing.