Suppress product open graph on pages without a product
On pages that feature a product, e.g. the product_info page, we want
to show the open graph meta information.  But on other pages, that
is unnecessary (and it's not clear what product's information we'd
show).

This tests that any `$product` that may exist has a get method.

We may revisit this in the future, but this makes things work with
the current state of the code.