Add hook to order builders
The database order builder loads the entire row from the orders table
but never allows anything to access it.  This modifies it by adding a
method to get the data and a hook to expose the builder itself.

Adds a similar hook to the cart order builder, but it's not clear
that that does anything.  Because the cart order builder gets its
information from the cart, which is already exposed.  It doesn't
have its own information.  This does expose the order object at build
time, which may be useful.